pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "expbox"
version = "0.1.0"
description = "A lightweight experiment context & result management helper."
authors = [
  {name = "Tadahaya Mizuno", email = "tadahaya@gmail.com"}
]
requires-python = ">=3.9"
dependencies = []

[project.optional-dependencies]
wandb = ["wandb>=0.16.0"]
yaml = ["pyyaml>=6.0"]

[project.urls]
Homepage = "https://github.com/mizuno-group/expbox"